Add new methods

+++ hrms/src/main/java/com/aberimen/hrms/JobPosition/JobPositionRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobPositionRepository extends JpaRepository<JobPosition, Long>{
+public interface JobPositionRepository extends JpaRepository<JobPosition, Integer>{
 
 	Optional<JobPosition> findByPositionNameIgnoreCase(String positionName);
